# Conflicts:
#	src/python/pants/backend/python/tasks/python_task.py

+++ src/python/pants/util/dirutil.py
@@ -19,25 +19,45 @@ from contextlib import contextmanager
 from pants.util.strutil import ensure_text
 
 
+def longest_dir_prefix(path, prefixes):
+  """Given a list of prefixes, return the one that is the longest prefix to the given path.
+
+  Returns None if there are no matches.
+  """
+  longest_match, longest_prefix = 0, None
+  for prefix in prefixes:
+    if fast_relpath_optional(path, prefix) is not None and len(prefix) > longest_match:
+      longest_match, longest_prefix = len(prefix), prefix
+
+  return longest_prefix
+
+
 def fast_relpath(path, start):
   """A prefix-based relpath, with no normalization or support for returning `..`."""
-  if not path.startswith(start):
-    raise ValueError('{} is not a prefix of {}'.format(start, path))
+  relpath = fast_relpath_optional(path, start)
+  if relpath is None:
+    raise ValueError('{} is not a directory containing {}'.format(start, path))
+  return relpath
 
-  if len(path) == len(start):
-    # Items are identical: the relative path is empty.
-    return ''
-  elif len(start) == 0:
+
+def fast_relpath_optional(path, start):
+  """A prefix-based relpath, with no normalization or support for returning `..`.
+
+  Returns None if `start` is not a directory-aware prefix of `path`.
+  """
+  if len(start) == 0:
     # Empty prefix.
     return path
-  elif start[-1] == '/':
-    # The prefix indicates that it is a directory.
-    return path[len(start):]
-  elif path[len(start)] == '/':
-    # The suffix indicates that the prefix is a directory.
-    return path[len(start)+1:]
-  else:
-    raise ValueError('{} is not a directory containing {}'.format(start, path))
+
+  # Determine where the matchable prefix ends.
+  pref_end = len(start) - 1 if start[-1] == '/' else len(start)
+  if pref_end > len(path):
+    # The prefix is too long to match.
+    return None
+  elif path[:pref_end] == start[:pref_end] and (len(path) == pref_end or path[pref_end] == '/'):
+    # The prefix matches, and the entries are either identical, or the suffix indicates that
+    # the prefix is a directory.
+    return path[pref_end+1:]
 
 
 def safe_mkdir(directory, clean=False):
@@ -290,6 +310,10 @@ def relative_symlink(source_path, link_path):
     raise ValueError("Path for link:{} must be absolute".format(link_path))
   if source_path == link_path:
     raise ValueError("Path for link is identical to source:{}".format(source_path))
+  # The failure state below had a long life as an uncaught error. No behavior was changed here, it just adds a catch.
+  # Raising an exception does differ from absolute_symlink, which takes the liberty of deleting existing directories.
+  if os.path.isdir(link_path) and not os.path.islink(link_path):
+    raise ValueError("Path for link would overwrite an existing directory: {}".format(link_path))
   try:
     if os.path.lexists(link_path):
       os.unlink(link_path)
